+++ Assets/Scripts/Weapon.cs
@@ -6,13 +6,13 @@ public class Weapon : ScriptableObject
     [SerializeField] GameObject _weaponPrefab;
     [SerializeField] GameObject _projectilePrefab;
 
-    [SerializeField] float _weaponDamage;
+    [SerializeField] int _weaponDamage;
     [SerializeField] float _fireRate;
     [SerializeField] float _projectileSpeed;
 
     private GameObject weapon;
 
-    public float GetDamage { get { return _weaponDamage; } }
+    public int GetDamage { get { return _weaponDamage; } }
     public float GetSpeed { get { return _projectileSpeed; } }
     public float GetFireRate { get { return 1/_fireRate; } }
